Changed default size of sample to float. This should be customizable later of course

+++ inc/mm_sample.h
@@ -1,6 +1,6 @@
 #ifndef MM_SAMPLE_H
 #define MM_SAMPLE_H 
 
-typedef double MMSample;
+typedef float MMSample;
 
 #endif /* MM_SAMPLE_H */
